Add mono configuration for publishing
Add a `mono.yml` configuration file so that it works with Mono.
https://github.com/appsignal/mono/

Use changesets to generate changesets and bump the version number.
See the Mono README for more information.

I changed the committed version down to 0.1.0 and added a changeset for
a major version bump so the next version will be version 1.0.0 on npm.

Closes #33