Update the publish workflow to listen to tags
If we keep the publish workflow as is it won't be triggers as one
GitHub Actions workflow can't trigger another workflow through an event
like creating a release on the repo, like I did in the previous commit.

Update the publish workflow to also listen to the tag push event.

I've also updated the `latest` tag push logic with the help of Claude to
figure out a way to not publish the `latest` tag when a release is a
prerelease.